Gravity has been added (not stable yet)

+++ include/Entities/CelestialBody.hpp
@@ -7,14 +7,14 @@
 class CelestialBody {
 protected:
 	float radius;
-	double mass;
+	float mass;
 	float x;
 	float y;
 	sf::Vector2f realV;  // Velocity
 	sf::Color color;
 
 public:
-	CelestialBody(float radius, double mass, float x, float y, sf::Vector2f realV, sf::Color color) : 
+	CelestialBody(float radius, float mass, float x, float y, sf::Vector2f realV, sf::Color color) : 
 		radius(radius),
 		mass(mass),
 		x(x),
@@ -27,14 +27,14 @@ public:
 	virtual void update() = 0;
 
 	const float getRadius() const;
-	const uint64_t getMass() const;
+	const float getMass() const;
 	const float getX() const;
 	const float getY() const;
 	const sf::Vector2f getRealV() const;
 	const sf::Color getColor() const;
 
 	void setRadius(float mass);
-	void setMass(uint64_t mass);
+	void setMass(float mass);
 	void setX(float x);
 	void setY(float y);
 	void setRealV(sf::Vector2f realV);
